+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## v1.4.0
 
-- Enhanced session configuration (new options for `sessionCookieKey`, `sessionHttpOnly`, `sessionMaxAge`, `sessionOverwrite`, `sessionRenew`, `sessionRolling` and `sessionSigned`)
+- Enhanced `bodyParser` configuration (new options for encoding, and extending with new types, disabling, setting size limits for form, JSON and text request body types)
+- Enhanced `session` configuration (new options for `sessionCookieKey`, `sessionHttpOnly`, `sessionMaxAge`, `sessionOverwrite`, `sessionRenew`, `sessionRolling` and `sessionSigned`)
 - Minor code improvements
 
 ## v1.3.3